requirements.txt
telethon
python-binance
python-dotenv